chore(context): extract types
